update classes OrderService and OrderServiceImpl

+++ trunk/stis/src/main/java/cz/muni/fi/pa165/stis/service/impl/OrderServiceImpl.java
@@ -1,6 +1,7 @@
 package cz.muni.fi.pa165.stis.service.impl;
 
 import cz.muni.fi.pa165.stis.dao.OrderDAO;
+import cz.muni.fi.pa165.stis.dto.CustomerTO;
 import cz.muni.fi.pa165.stis.dto.OrderTO;
 import cz.muni.fi.pa165.stis.entity.Customer;
 import cz.muni.fi.pa165.stis.entity.Order;
@@ -27,23 +28,44 @@ public class OrderServiceImpl implements OrderService{
 
     @Transactional
     public void create(OrderTO order) {
+        if (order == null) {
+            throw new IllegalArgumentException("order is null");
+        }
+        if (order.getId() != null) {
+            throw new IllegalArgumentException("order.id is not null");
+        }
         Order ord = mapper.map(order, Order.class);
         orderDAO.create(ord);
     }
 
     @Transactional(readOnly=true)
     public OrderTO get(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("id is null");
+        }
         return mapper.map(orderDAO.get(id), OrderTO.class);
     }
 
     @Transactional
     public void update(OrderTO order) {
+        if (order == null) {
+            throw new IllegalArgumentException("order is null");
+        }
+        if (order.getId() != null) {
+            throw new IllegalArgumentException("order.id is not null");
+        }
         Order ord = mapper.map(order, Order.class);
         orderDAO.update(ord);
     }
 
     @Transactional
     public void remove(OrderTO order) {
+        if (order == null) {
+            throw new IllegalArgumentException("order is null");
+        }
+        if (order.getId() != null) {
+            throw new IllegalArgumentException("order.id is not null");
+        }
         Order ord = mapper.map(order, Order.class);
         orderDAO.remove(ord);
     }
@@ -59,8 +81,11 @@ public class OrderServiceImpl implements OrderService{
     }
 
     @Transactional(readOnly=true)
-    public List<OrderTO> findByCustomer(Customer customer) {
-        List<Order> result = orderDAO.findByCustomer(customer);
+    public List<OrderTO> findByCustomer(CustomerTO customer) {
+        if (customer == null) {
+            throw new IllegalArgumentException("customer is null");
+        }
+        List<Order> result = orderDAO.findByCustomer(mapper.map(customer, Customer.class));
         List<OrderTO> ret = new ArrayList<OrderTO>();
         for (Order ord : result) {
             ret.add(mapper.map(ord, OrderTO.class));
